+++ tests/integration/collection_version/updates/test/add_field_test.go
