+++ tests/integration/collection_version/updates/test/name_test.go
@@ -17,7 +17,7 @@ import (
 	testUtils "github.com/sourcenetwork/defradb/tests/integration"
 )
 
-func TestColVersionUpdateTestName(t *testing.T) {
+func TestColVersionUpdateTestNameByVersionID(t *testing.T) {
 	test := testUtils.TestCase{
 		Actions: []any{
 			&action.AddSchema{
@@ -42,7 +42,7 @@ func TestColVersionUpdateTestName(t *testing.T) {
 	testUtils.ExecuteTestCase(t, test)
 }
 
-func TestColVersionUpdateTestName_Fails(t *testing.T) {
+func TestColVersionUpdateTestNameByVersionID_Fails(t *testing.T) {
 	test := testUtils.TestCase{
 		Actions: []any{
 			&action.AddSchema{
@@ -65,5 +65,50 @@ func TestColVersionUpdateTestName_Fails(t *testing.T) {
 		},
 	}
 
+	testUtils.ExecuteTestCase(t, test)
+}
+
+func TestColVersionUpdateTestName(t *testing.T) {
+	test := testUtils.TestCase{
+		Actions: []any{
+			&action.AddSchema{
+				Schema: `
+					type Users {
+						name: String
+					}
+				`,
+			},
+			testUtils.PatchCollection{
+				Patch: `
+					[
+						{ "op": "test", "path": "/Users/Name", "value": "Users" }
+					]
+				`,
+			},
+		},
+	}
+	testUtils.ExecuteTestCase(t, test)
+}
+
+func TestColVersionUpdateTestName_Fails(t *testing.T) {
+	test := testUtils.TestCase{
+		Actions: []any{
+			&action.AddSchema{
+				Schema: `
+					type Users {
+						name: String
+					}
+				`,
+			},
+			testUtils.PatchCollection{
+				Patch: `
+					[
+						{ "op": "test", "path": "/Users/Name", "value": "Book" }
+					]
+				`,
+				ExpectedError: "failed: test failed",
+			},
+		},
+	}
 	testUtils.ExecuteTestCase(t, test)
 }
